third_party/WebKit/LayoutTests/fast/events/gesture-pinch-zoom-scroll-bubble-expected.txt

PASS internals.visualViewportWidth() is 400
PASS internals.visualViewportHeight() is 300
PASS internals.visualViewportScrollX() is 200
PASS internals.visualViewportScrollY() is 300
PASS internals.visualViewportScrollX() is 200
PASS internals.visualViewportScrollY() is 301
PASS internals.visualViewportScrollX() is 201
PASS internals.visualViewportScrollY() is 301
PASS internals.visualViewportScrollX() is 199
PASS internals.visualViewportScrollY() is 299
PASS successfullyParsed is true

TEST COMPLETE
